Publish beside a waiting bus receive, report a failed link receive, and rewrite the bus and link guides
Event bus:
- EventPublisher: a publish-only handle that clones into every part that
  announces and returns how many subscribers it reached.
- In every binding an endpoint's publisher sits apart from its receiver, so
  a publish or a subscribe no longer waits behind an open next. In C# they
  no longer reach the native endpoint alongside one.
- An endpoint counts what it missed by falling behind, and TypeScript and
  C# get a wait with a time limit.
- The capacity is documented as rounded up to a power of two and capped at
  MAX_CAPACITY. A negative capacity or time limit in TypeScript is refused
  rather than wrapped to a huge unsigned value.

Your own link:
- A link's recv that fails is reported by the next receive in every
  binding and through the C ABI, rather than ending the link silently.
- A TypeScript handler that throws before returning fails the call rather
  than the Node process.
- A delivered payload may be text in every language.

Both guides print the same eight lines in all four languages, with tables
of the calls and values and what each error means.

+++ bindings/node/src/coap.rs
@@ -96,7 +96,7 @@ impl CoapClient {
     /// @param timeoutMs - how long to wait before rejecting; a message that arrives later
     /// waits for the next receive.
     #[napi]
-    pub async fn recv(&self, timeout_ms: Option<u32>) -> napi::Result<Option<TransportMessage>> {
+    pub async fn recv(&self, timeout_ms: Option<f64>) -> napi::Result<Option<TransportMessage>> {
         let inner = Arc::clone(&self.inner);
         within(timeout_ms, async move {
             let mut transport = inner.lock().await;
@@ -184,7 +184,7 @@ impl CoapServer {
     /// @param timeoutMs - how long to wait before rejecting; a reading that arrives
     /// later waits for the next receive.
     #[napi]
-    pub async fn recv(&self, timeout_ms: Option<u32>) -> napi::Result<Option<TransportMessage>> {
+    pub async fn recv(&self, timeout_ms: Option<f64>) -> napi::Result<Option<TransportMessage>> {
         let inner = Arc::clone(&self.server);
         within(timeout_ms, async move {
             let mut server = inner.lock().await;
